fix: apply the indexer's ReDoS gate to editor tag highlighting
The content script compiled user regexes without isRegexSafe, which the
indexer has applied since it was written. There was no reason for the
divergence; the editor is the path that needed it more.

The two surfaces face different hazards and each was guarded for only one:

- the indexer matches with String.prototype.match, which advances past
  zero-length matches per spec ('a #tag b'.match(/\b\w*/g) terminates), so it
  never needed the empty-match guard
- the editor drives MatchDecorator's exec loop, so it needed that guard, but
  had no backtracking gate at all - and it runs the pattern over the viewport
  on every keystroke, on the main thread, where the indexer runs in batches
  off the interaction path

Measured with (?<=^|\s)#(\w+|\w+)*$, which isRegexSafe rejects: a single exec
against an 18-character line takes ~10s, and 22 and 26 characters did not
return within 110s. That was reachable from an advanced setting.

Converging costs nothing documented: the default, the multi-prefix example
from the Tag regex description, the capture-group form, plain #[a-z]+ and the
hex-colour exclude example all pass. A pattern isRegexSafe rejects was already
being ignored by the indexer, which fell back to the default, so the editor was
painting tags that search did not honour.

isRegexSafe moves to utils.ts unchanged so both paths share it. Verified by
defect injection: disabling the gate fails 2 of the new tests.
